Fix curl error handling and format error string
This also gives a hint for curl_error() == 7 which normally leads to SELinux.
Test that on a box with SELinux in enforcing mode.

Also fixed error handling of getPreviewImage (renamed to getPreviewHtml).
This commit passed the previewHtml variable by reference while returning true/false.
This allows to check whether to render the URL or not.

Simplified the code in some parts too.

refs #33
fixes #35