Render html before the image
With this change, the graph url is generated and returned to the browser as a
base64 string inside the imgsrc. Now the user doesn't have to wait the image
from grafana to view information about service/host, the image is sent
in a separated http connection.

The image is downloaded and sent to browser when init function is called (before
any header or html is sent by Icinga framework).