Found Difference in provided dashboards on import
On the initial import, I found that content is manually needed for this dashboard in order for Grafana to accept the import. For `icinga2-default.json`, "null" is specified and one is able to import the dashboard without any immediate changes. Requesting to update the first dashboard file to match the other for consistency.