Update Grapher.php
Hi Carsten
We have the problem, that the grafana-renderer is very slow and uses a lot resources. Is much faster, to use iframes to embed graphs in icingaweb2. Further you then have tooltips in the graphs.
The downside is, that you do no longer have links to grafana to open a dashboard from icingaweb2.
So I extended Grapher.php to add a link above the grafana panels embedded as iframes into icingaweb2. With this, you have much faster rendering of the graphs in the iframes and you still can click on a link '=> see in Grafana' to open the panel in Grafana.

I did the according changes to GeneralConfigForm.php to be able to interactively do the config for the grafana-module too.